requirements.txt
Flask-Login==0.6.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
requests==2.31.0
bcrypt==4.0.1